Finishing
Finishing my coding up. I have now written very in-depth comments explaining why I made the decisions I did in the different sections of my code. I also write about what the code is doing and its effect on the rest.

I chose not to write one for the JSON file, since it is pretty much as you see it. You assign each product an ID and provide the properties' key values.

So now it all has an explanation, and you know why I didn't do it for JSON.